requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aiohttp==1.0.5
async-timeout==1.2.1      # via aiohttp
asyncio==3.4.3
chardet==3.0.2            # via aiohttp
discord.py==0.16.7
multidict==2.1.4          # via aiohttp
pylast==1.8.0
six==1.10.0               # via pylast
websockets==3.3           # via discord.py
